+++ src/utils/getFiles.js
@@ -11,26 +11,49 @@ var path = require( 'path' )
  * @returns {Array} returns an array of files
 */
 var getFiles = function( dir ) {
-	if ( typeof dir !== 'string' ) {
-		throw new TypeError( 'getFiles err. Expected string, but received: ' + typeof dir )
+	if ( typeof dir !== 'string' && !( dir instanceof Array ) ) {
+		throw new TypeError( 'getFiles err. Expected string or array, but received: ' + typeof dir )
 	}
 
-	glob( dir, {}, function( err, files ) {
-		if ( err ) { throw err }
-		files = files.filter( function( file ) {
+	if ( typeof dir === 'string' ) {
+		return glob( dir, {}, function( err, files ) {
+			if ( err ) { throw err }
+
+			files = files.filter( function( file ) {
+				var excluded = false
+				var relPath = path.relative( dir.replace( '/**/*.styl', '' ), file )
+
+				this.config.exclude.forEach( function( exclude ) {
+					excluded = excluded || exclude.match( relPath )
+				} )
+
+				return !excluded
+			}, this )
+
+			this.cache.filesLen = files.length - 1
+			this.cache.files = files
+
+			return async.map( this.cache.files, fs.readFile, this.parse.bind( this ) )
+		}.bind( this ) )
+	}
+	else if ( dir instanceof Array ) {
+
+		var files = dir.filter( function( filepath ) {
 			var excluded = false
-			var relPath = path.relative( dir.replace( '/**/*.styl', '' ), file )
 
 			this.config.exclude.forEach( function( exclude ) {
-				excluded = excluded || exclude.match( relPath )
+				excluded = excluded || exclude.match( filepath )
 			} )
 
 			return !excluded
 		}, this )
+
 		this.cache.filesLen = files.length - 1
 		this.cache.files = files
-		return async.map( this.cache.files, fs.readFile, this.parse.bind( this ) )
-	}.bind( this ) )
+		return this.cache.files.forEach( function( file ) {
+			return this.read( file )
+		}.bind( this ) )
+	}
 }
 
 module.exports = getFiles